refactor eval script into smaller files and use relative imports for clean code
